Moved entry point and all admin code to new app.py
